requirements.txt
asgiref==3.6.0
Django==4.1.7
django-environ==0.10.0
django-filter==22.1
djangorestframework==3.14.0
importlib-metadata==6.0.0
Markdown==3.4.1
model-bakery==1.11.0
mysqlclient==2.1.1
Pillow==9.5.0
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7
zipp==3.15.0
